+++ workers/ohsome_quality_analyst/api/request_models.py
@@ -10,6 +10,7 @@ information derived from `pydantic` models in the automatic generated API docume
 import json
 from enum import Enum
 
+
 import geojson
 import pydantic
 from geojson import Feature, FeatureCollection
@@ -114,19 +115,7 @@ class BaseBpolys(BaseModel):
         return obj
 
 
-class BaseDatabase(BaseModel):
-    """Model for the combination of parameters: `dataset`, `feature_id`, `fid_field`."""
-
-    dataset: DatasetEnum = pydantic.Field(..., title="Dataset Name", example="regions")
-    feature_id: str = pydantic.Field(..., title="Feature Id", example="3")
-    fid_field: FidFieldEnum | None = None
-
-
 class IndicatorBpolys(BaseIndicator, BaseTopicName, BaseBpolys):
-    pass
-
-
-class IndicatorDatabase(BaseIndicator, BaseTopicName, BaseDatabase):
     pass
 
 
@@ -138,28 +127,9 @@ class ReportBpolys(BaseReport, BaseBpolys):
     pass
 
 
-class ReportDatabase(BaseReport, BaseDatabase):
-    pass
-
-
 INDICATOR_EXAMPLES = {
-    "OQT AOI": {
-        "summary": (
-            "Request an Indicator for an AOI defined by OQT (`dataset` and "
-            "`featureId`)."
-        ),
-        "value": {
-            "topic": "building-count",
-            "dataset": "regions",
-            "featureId": 3,
-            "fidField": "ogc_fid",
-            "includeSvg": False,
-            "includeHtml": False,
-            "flatten": False,
-        },
-    },
-    "Custom AOI": {
-        "summary": "Request an Indicator for a custom AOI (`bpolys`).",
+    "AOI": {
+        "summary": "Request an Indicator for a given AOI (`bpolys`).",
         "value": {
             "topic": "building-count",
             "bpolys": {
@@ -177,117 +147,6 @@ INDICATOR_EXAMPLES = {
                     ],
                 },
             },
-        },
-    },
-    "Custom AOI and custom Topic": {
-        "summary": (
-            "Request an Indicator for a custom AOI (`bpolys`) and a custom Topic "
-            "(`topic`)."
-        ),
-        "value": {
-            "bpolys": {
-                "type": "Feature",
-                "geometry": {
-                    "type": "Polygon",
-                    "coordinates": [
-                        [
-                            [8.674092292785645, 49.40427147224242],
-                            [8.695850372314453, 49.40427147224242],
-                            [8.695850372314453, 49.415552187316095],
-                            [8.674092292785645, 49.415552187316095],
-                            [8.674092292785645, 49.40427147224242],
-                        ]
-                    ],
-                },
-            },
-            "topic": {
-                "key": "my-topic-key",
-                "name": "My topic name",
-                "description": "My topic description",
-                "data": {
-                    "result": [
-                        {"timestamp": "2014-01-01T00:00:00Z", "value": 4708},
-                        {"timestamp": "2014-02-01T00:00:00Z", "value": 4842},
-                        {"timestamp": "2014-03-01T00:00:00Z", "value": 4840},
-                        {"timestamp": "2014-04-01T00:00:00Z", "value": 4941},
-                        {"timestamp": "2014-05-01T00:00:00Z", "value": 4987},
-                        {"timestamp": "2014-06-01T00:00:00Z", "value": 5007},
-                        {"timestamp": "2014-07-01T00:00:00Z", "value": 5020},
-                        {"timestamp": "2014-08-01T00:00:00Z", "value": 5168},
-                        {"timestamp": "2014-09-01T00:00:00Z", "value": 5355},
-                        {"timestamp": "2014-10-01T00:00:00Z", "value": 5394},
-                        {"timestamp": "2014-11-01T00:00:00Z", "value": 5449},
-                        {"timestamp": "2014-12-01T00:00:00Z", "value": 5470},
-                        {"timestamp": "2015-01-01T00:00:00Z", "value": 5475},
-                        {"timestamp": "2015-02-01T00:00:00Z", "value": 5477},
-                        {"timestamp": "2015-03-01T00:00:00Z", "value": 5481},
-                        {"timestamp": "2015-04-01T00:00:00Z", "value": 5495},
-                        {"timestamp": "2015-05-01T00:00:00Z", "value": 5516},
-                        {"timestamp": "2015-06-01T00:00:00Z", "value": 5517},
-                        {"timestamp": "2015-07-01T00:00:00Z", "value": 5519},
-                        {"timestamp": "2015-08-01T00:00:00Z", "value": 5525},
-                        {"timestamp": "2015-09-01T00:00:00Z", "value": 5560},
-                        {"timestamp": "2015-10-01T00:00:00Z", "value": 5564},
-                        {"timestamp": "2015-11-01T00:00:00Z", "value": 5568},
-                        {"timestamp": "2015-12-01T00:00:00Z", "value": 5627},
-                        {"timestamp": "2016-01-01T00:00:00Z", "value": 5643},
-                        {"timestamp": "2016-02-01T00:00:00Z", "value": 5680},
-                        {"timestamp": "2016-03-01T00:00:00Z", "value": 5681},
-                        {"timestamp": "2016-04-01T00:00:00Z", "value": 5828},
-                        {"timestamp": "2016-05-01T00:00:00Z", "value": 5974},
-                        {"timestamp": "2016-06-01T00:00:00Z", "value": 5990},
-                        {"timestamp": "2016-07-01T00:00:00Z", "value": 5991},
-                        {"timestamp": "2016-08-01T00:00:00Z", "value": 5997},
-                        {"timestamp": "2016-09-01T00:00:00Z", "value": 6002},
-                        {"timestamp": "2016-10-01T00:00:00Z", "value": 6010},
-                        {"timestamp": "2016-11-01T00:00:00Z", "value": 6010},
-                        {"timestamp": "2016-12-01T00:00:00Z", "value": 6016},
-                        {"timestamp": "2017-01-01T00:00:00Z", "value": 6015},
-                    ]
-                },
-            },
-            "includeSvg": False,
-            "includeHtml": False,
-            "flatten": False,
-        },
-    },
-}
-
-REPORT_EXAMPLES = {
-    "OQT AOI": {
-        "summary": (
-            "Request a Report for a AOI defined by OQT (`dataset` and `featureId`)."
-        ),
-        "value": {
-            "dataset": "regions",
-            "featureId": 12,
-            "fidField": "ogc_fid",
-            "includeSvg": False,
-            "includeHtml": False,
-            "flatten": False,
-        },
-    },
-    "Custom AOI": {
-        "summary": "Request a Report for a custom AOI (`bpolys`).",
-        "value": {
-            "bpolys": {
-                "type": "Feature",
-                "geometry": {
-                    "type": "Polygon",
-                    "coordinates": [
-                        [
-                            [8.674092292785645, 49.40427147224242],
-                            [8.695850372314453, 49.40427147224242],
-                            [8.695850372314453, 49.415552187316095],
-                            [8.674092292785645, 49.415552187316095],
-                            [8.674092292785645, 49.40427147224242],
-                        ]
-                    ],
-                },
-            },
-            "includeSvg": False,
-            "includeHtml": False,
-            "flatten": False,
         },
     },
 }