+++ workers/tests/integrationtests/api/test_indicator_geojson_io.py
@@ -2,7 +2,6 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 """
-import os
 import unittest
 from datetime import datetime, timedelta
 
@@ -72,26 +71,6 @@ class TestApiIndicatorIo(unittest.TestCase):
         self.assertEqual(response.status_code, 422)
         content = response.json()
         self.assertEqual(content["type"], "SizeRestrictionError")
-
-    @oqt_vcr.use_cassette()
-    def test_invalid_set_of_arguments(self):
-        path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "..",
-            "fixtures",
-            "heidelberg-altstadt-feature.geojson",
-        )
-        with open(path, "r") as f:
-            bpolys = f.read()
-        parameters = {
-            "bpolys": bpolys,
-            "dataset": "foo",
-            "feature_id": "3",
-        }
-        response = self.client.post(self.endpoint, json=parameters)
-        self.assertEqual(response.status_code, 422)
-        content = response.json()
-        self.assertEqual(content["type"], "RequestValidationError")
 
     @oqt_vcr.use_cassette()
     def test_indicator_include_svg(self):