+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## Current Main
 
+### Breaking
+
+- api: remove `dataset`, `featureId` and `fidField` parameter ([#634])
+- cli: remove CLI ([#634])
+
 ### Other Changes
 
-- style: use `ruff` instead `flake8` and `isort`
+- style: use `ruff` instead `flake8` and `isort` ([#630])
+
+[#630](https://github.com/GIScience/ohsome-quality-analyst/pull/630)
+[#634](https://github.com/GIScience/ohsome-quality-analyst/pull/634)
 
 ## 0.15.0